fix linux/{386,arm} issues

+++ atomic.go
@@ -40,7 +40,7 @@ func a_store_16(addr uintptr, val uint16) {
 
 // static inline int a_ctz_l(unsigned long x)
 func _a_ctz_l(tls *TLS, x ulong) int32 {
-	return int32(mbits.TrailingZeros64(x))
+	return int32(mbits.TrailingZeros64(uint64(x)))
 }
 
 // static inline int a_ctz_64(uint64_t x)
